requirements.txt
streamlit==1.7.0
numpy==1.22.3
mtcnn==0.1.1
Pillow==9.0.1